+++ src/baseconfig.h
@@ -25,6 +25,9 @@ class BaseConfig {
     const uint8_t&  GetDebugLevel()    const {return debuglevel;}
     const uint8_t&  GetSerialRx()      const {return serial_rx;}
     const uint8_t&  GetSerialTx()      const {return serial_tx;}
+    const bool&     GetUseAuth()        const { return useAuth; }
+    const String&   GetAuthUser()      const {return auth_user;}
+    const String&   GetAuthPass()      const {return auth_pass;}
 
     const String    GetReleaseName(); 
   private:
@@ -40,6 +43,9 @@ class BaseConfig {
     uint8_t   debuglevel;
     uint8_t   serial_rx;
     uint8_t   serial_tx;
+    bool      useAuth;
+    String    auth_user;
+    String    auth_pass;
 
 };